CFS_Iberia_hus850 {transformeR}	R Documentation

Multimember grid containing CFSv2 seasonal forecast data of daily specific humidity at 850mb for the Iberian Peninsula (DJF, 1983-2002).

Description

This R data object was obtained from the ECOMS-UDG server (http://www.meteo.unican.es/ecoms-udg, log in is requiered, see loginUDG) by means of function loadECOMS (package loadeR.ECOMS) in the following manner:

years <- 1983:2002
season <- c(12,1,2)
lonLim <- c(-10, 4)
latLim <- c(35, 44)

loginUDG("username", "pasword")

CFS_Iberia_hus@850 <- loadECOMS(dataset = "CFSv2_seasonal", var = "huss@850", years = years, lonLim = lonLim, latLim = latLim, season = season, time = "DD", aggr.d = "mean", leadMonth = 1, members = 1:9)

Format

A multimember grid

Source

http://www.meteo.unican.es/ecoms-udg

References

Saha, S. et al., 2013. The NCEP Climate Forecast System Version 2. J Clim 130925135638001. doi:10.1175/JCLI-D-12-00823.1

Examples

data(CFS_Iberia_hus850)
plotClimatology(climatology(CFS_Iberia_hus850, clim.fun = list(FUN = sum)),
                backdrop.theme = "coastline")

[Package transformeR version 1.4.2 Index]