EOBS_Iberia_pr {transformeR}	R Documentation

Grid containing E-OBS daily data of precipitation for the Iberian Peninsula (DJF, 1983-2002).

Description

This R data object was obtained from the UDG server (http://www.meteo.unican.es/udg-tap, log in is requiered, see loginUDG) by means of function loadGridData (package loadeR) in the following manner:

years <- 1983:2002
season <- c(12,1,2)
lonLim <- c(-10, 4)
latLim <- c(35, 44)

loginUDG("username", "pasword")

EOBS_Iberia_pr <- loadGridData(dataset = "http://meteo.unican.es/tds5/dodsC/eobs/e-obs_v14_0.50regular.ncml", var = "rr", years = years, lonLim = lonLim, latLim = latLim, season = season)

Format

A grid

Source

http://www.meteo.unican.es/udg-tap

See Also

makeMultiGrid, loadGridData

Examples

data("EOBS_Iberia_pr")
total.precip.annual <- aggregateGrid(EOBS_Iberia_pr,
                                     aggr.m = list(FUN = sum),
                                     aggr.y = list(FUN = sum))
plotClimatology(climatology(total.precip.annual), 
                backdrop.theme = "countries", scales = list(draw = TRUE),
                main = "Mean Total DJF precip (mm) 1998-2000")

[Package transformeR version 1.4.2 Index]