VALUE_Iberia_pr {transformeR}	R Documentation

Station data containing VALUE (ECA-D) daily data of precipitation for the Iberian Peninsula (DJF, 1983-2002).

Description

This R data object was obtained by means of function loadStationData (package loadeR) in the following manner:

years <- 1983:2002
season <- c(12,1,2)
lonLim <- c(-10, 4)
latLim <- c(35, 44)

VALUE_Iberia_pr <- loadStationData(dataset = "data/VALUE_ECA_86_v2.zip", var = "precip", years = years, lonLim = lonLim, latLim = latLim, season = season)

Station data from the VALUE (ECA-D) dataset containing daily precipitation for stations in the Iberian Peninsula.

Format

Station data

Source

http://www.meteo.unican.es/udg-tap

Subset of VALUE (ECA-D) station data. Full dataset is accessible for download in http://meteo.unican.es/work/downscaler/data/VALUE_ECA_86_v2.tar.gz.

See Also

makeMultiGrid, loadGridData

Examples

data("VALUE_Iberia_pr")
plotClimatology(climatology(VALUE_Iberia_pr),
                backdrop.theme = "countries", scales = list(draw = TRUE))

[Package transformeR version 1.4.2 Index]