bindGrid.time {transformeR}	R Documentation

Grid binding by time dimension

Description

Flexible binding of (spatially/member consistent) grids by their time dimension

Usage

bindGrid.time(..., spatial.tolerance = 0.001)

Arguments

...
Input grids to bind by their time dimension. These must be compatible in members and space (see details). For flexibility, they can be introduced as a list or directly as consecutive arguments.

spatial.tolerance
numeric. Coordinate differences smaller than spatial.tolerance will be considered equal coordinates. Default to 0.001 –assuming that degrees are being used it seems a reasonable rounding error after interpolation–. This value is passed to the identical function to check for spatial consistency of the input grids.

Details

Input grids consistency checks

The function makes a number of checks in order to test the spatiotemporal compatibility of the input multi-member grids. The spatial consistency of the input grids is also checked. In order to avoid possible errors from the user, the spatial consistency (i.e., equal XY coordinates) of the input grids must be ensured before attempting the creation of the multigrid, otherwise giving an error. This can be achieved either through the specification of the same 'lonLim' and 'latLim' argument values when loading the grids, or using the interpGrid interpolator in conjuntion with the getGrid method, for instance

Time dimension ordering

The "time" dimension will be always returned in ascending order, regardless of the ordering of the input grids. The internal helper sortDim.time is applied to this aim.

Author(s)

M De Felice, J Bedia

See Also

Other internal.helpers: array3Dto2Dmat.stations, array3Dto2Dmat, bindGrid.member, bindGrid.spatial, bindGrid, grid3Dto2Dmatrix, mat2Dto3Darray.stations, redim, sortDim.spatial, sortDim.time

Examples

data("EOBS_Iberia_tas")
eobs.1998 <- subsetGrid(EOBS_Iberia_tas, years = 1998)
eobs.1999_2000 <- subsetGrid(EOBS_Iberia_tas, years = 1999:2000)
eobs1 <- bindGrid.time(eobs.1998, eobs.1999_2000)
eobs2 <- bindGrid.time(eobs.1999_2000, eobs.1998)
identical(eobs1, eobs2)

# For convenience while programming, it also accepts a list as input:
year.list <- list(1998, 1999:2000)
grid.list <- lapply(year.list, function(x) {
    subsetGrid(EOBS_Iberia_tas, years = x)
    # perform any other tasks with the subsets...
}) 
eobs3 <- do.call("bindGrid.time", grid.list)
identical(eobs1, eobs3)

data("CFS_Iberia_tas")
# We first diaggregate in various grids with different time periods
period1 <- subsetGrid(CFS_Iberia_tas, years = 1983:1991)
period2 <- subsetGrid(CFS_Iberia_tas, years = 1992:2002)
# Then we aggregate and compare to the original data (containing the full continuous period)
bindedGrid <- bindGrid.time(period1, period2)
plotClimatology(climatology(bindedGrid), backdrop.theme = "coastline", at = seq(0,17,1))
plotClimatology(climatology(CFS_Iberia_tas), backdrop.theme = "coastline", at = seq(0,17,1))

[Package transformeR version 1.4.2 Index]