clim2sgdf {transformeR}	R Documentation

Climatology to SpatialGridDataFrame or SpatialPointsDataFrame

Description

Convert a climatological grid to a SpatialGridDataFrame object from package sp

Usage

clim2sgdf(clim, set.min, set.max)

Arguments

clim
A climatological grid, as returned by function climatology

set.min
Minimum value, as passed by plotClimatology

set.max
Maximum value, as passed by plotClimatology

Details

This function is intended for internal usage by plotClimatology, that accepts all possible arguments that can be passed to spplot for plotting. However, it may be useful for advanced sp users in different contexts (e.g. for reprojecting via spTransform etc.)

Value

A sp object of the class SpatialGridDataFrame

Author(s)

J. Bedia

See Also

climatology, plotClimatology

Examples

data("CFS_Iberia_tas")
# Climatology is computed:
clim <- climatology(CFS_Iberia_tas, by.member = TRUE)
sgdf <- clim2sgdf(clim, NULL, NULL)
class(sgdf)

[Package transformeR version 1.4.2 Index]