draw.world.lines {transformeR}	R Documentation

Land borders

Description

Add land borders to a map

Usage

draw.world.lines(...)

Arguments

...
Graphical parameters passed to lines.

Details

The function loads a built-in world segments dataset created ad hoc to avoid dependencies on other packages (i.e. 'maps'). Geographical lonlat coordinates in wgs84.

Value

Draws a simplied land border areas as lines onto the map

Author(s)

J. Bedia

Source

Postprocessed from the original shapefile from Natural Earth (http://www.naturalearthdata.com/downloads/110m-physical-vectors/)

[Package transformeR version 1.4.2 Index]