easyVeri.signif {transformeR}	R Documentation

Identify points significantly different from a given ref value

Description

Two-tailed hypothesis testing for easyVerification sigma outputs

Usage

easyVeri.signif(ref.mean, est.mean, sigma, cl = 0.95)

Arguments

ref.mean
Numeric of length one. Null hypothesis reference mean (e.g. 0.5 for AUC tests, 0 for skill scores)

est.mean
easyVerification output matrix with scores/measures. This is the cat[num] element of the veriApply output

sigma
Standard deviation of the estimated mean. This is the cat[num].sigma element of the veriApply output.

cl
Confidence level. Default to 0.95

Value

A easyVerification-like binary matrix (1 significant / 0 non-significant). Ready to be passed to easyVeri2grid for climatological grid conversion, and possibly to map.stippling after that, to depict significant points in a verification map.

Author(s)

J Bedia

See Also

easyVeri2grid, for conversion of verification matrices into climatological grids. Also, map.stippling, to draw significant points on verification maps.

Other easyVerification.bridges: easyVeri2grid

[Package transformeR version 1.4.2 Index]