easyVeri2grid {transformeR}	R Documentation

easyVerification matrix to climatological grid conversion

Description

Convert a xyz-type verification matrix as returned by veriApply to a climatological grid

Usage

easyVeri2grid(easyVeri.mat, obs.grid, verifun = NULL)

Arguments

easyVeri.mat
A matrix containing the verification measure, as returned by veriApply

obs.grid
The grid containing the verifying reference used in the call to veryApply producing the score matrix.

verifun
Optional. Character string indicating the value of the verifun argument. Just for a better traceability and metadata completeness.

Value

A climatological grid.

Author(s)

J. Bedia

See Also

climatology, plotClimatology.

Other easyVerification.bridges: easyVeri.signif

[Package transformeR version 1.4.2 Index]