filterNA {transformeR}	R Documentation

Remove missing values from grids

Description

Removes the dates for all grid points whenever there exists missing values such as NaN or NA.

Usage

filterNA(grid)

Arguments

grid
The input grid to be filtered.

Details

The function also takes care of adjusting dates and other relevant metadata (via the internal subsetDimension).

Value

A new grid object without missing values

Author(s)

J. Bano-Medina, J. Bedia

See Also

Other downscaling.helpers: convert2bin, makeMultiGrid

Examples

# Check if the dataset contains missing values (YES):
anyNA(VALUE_Iberia_pr$Data)
getShape(VALUE_Iberia_pr)
na.filtered <- filterNA(VALUE_Iberia_pr)
# Check if the dataset contains missing values (NO):
anyNA(na.filtered$Data)
getShape(na.filtered)

[Package transformeR version 1.4.2 Index]