getCoordinates {transformeR}	R Documentation

Get geographical coordinates of a climate data object

Description

Returns the coordinates of a climate data object, either stations or grid

Usage

getCoordinates(obj)

Arguments

obj
Any object extending the station or grid classes

Value

A list with x and y components

Author(s)

J. Bedia

See Also

Other get.helpers: get2DmatCoordinates, getDim, getGridVerticalLevels, getGrid, getRefDates, getSeason, getShape, getTimeResolution, getVarNames, getYearsAsINDEX, typeofGrid, which.leap

Examples

# Regular grid
data("NCEP_Iberia_hus850")
str(getCoordinates(NCEP_Iberia_hus850))
# Regular rotated grid with correpondence with a non regular grid in the 
# lat/lon domain.
data("CORDEX_Iberia_tas")
str(getCoordinates(CORDEX_Iberia_tas))
# Station data
data("VALUE_Iberia_tas")
str(getCoordinates(VALUE_Iberia_tas))

[Package transformeR version 1.4.2 Index]