intersectGrid {transformeR}	R Documentation

Intersection of multiple grids

Description

Takes multiple input grids and crops the overlapping part

Usage

intersectGrid(..., type = c("temporal", "spatial"), which.return = 1)

Arguments

...
Input grids

type
Character. Options are "temporal" (default) or "spatial".

which.return
Integer of the index to specify which grids in "..." are to be returned.

Value

The grids indicated in which.return, encompassing the overlapping time period.

Author(s)

M Iturbide

See Also

subsetGrid

Other subsetting: getTemporalIntersection, intersectGrid.spatial, intersectGrid.time, overGrid, subsetDimension, subsetGrid, subsetMembers, subsetRuntime, subsetSeason, subsetSpatial, subsetStation, subsetVar, subsetYears

Examples

{
data("EOBS_Iberia_tas")
a <- subsetGrid(EOBS_Iberia_tas, lonLim = c(-8,-1), latLim = c(37, 40))
b <- subsetGrid(EOBS_Iberia_tas, lonLim = c(-4,3), latLim = c(39, 43))
z <- intersectGrid(a, b, type = "spatial", which.return = 1)
}

[Package transformeR version 1.4.2 Index]