intersectGrid.time {transformeR}	R Documentation

Temporal intersection of multiple grids

Description

Takes multiple input grids and crops the overlapping part along time dimension

Usage

intersectGrid.time(..., which.return = 1)

Arguments

...
Input grids

which.return
Integer of the index to specify which grids in "..." are to be returned.

Value

The grids indicated in which.return, encompassing the overlapping time period.

Author(s)

M Iturbide

See Also

checkDim, checkSeason, getYearsAsINDEX, getSeason, for other time dimension helpers

Other subsetting: getTemporalIntersection, intersectGrid.spatial, intersectGrid, overGrid, subsetDimension, subsetGrid, subsetMembers, subsetRuntime, subsetSeason, subsetSpatial, subsetStation, subsetVar, subsetYears

[Package transformeR version 1.4.2 Index]