isRegular {transformeR}	R Documentation

Check if the input grid is regular

Description

Check if the coordinates in the data are regular or irregular

Usage

isRegular(grid)

Arguments

grid
a grid or station data, or the object returned by function getGrid

Value

Logical.

Author(s)

M Iturbide

See Also

getGrid

typeofGrid, for a character representation of the type of grid

Examples

data("EOBS_Iberia_tas")
isRegular(EOBS_Iberia_tas)
isRegular(getGrid(EOBS_Iberia_tas))
data("VALUE_Iberia_tas")
isRegular(VALUE_Iberia_tas)

[Package transformeR version 1.4.2 Index]