mat2Dto3Darray.stations {transformeR}	R Documentation

Conversion 2D matrix into a 3D array for station data

Description

Converts a 2D matrix of the form [time, lonlat] to a 3D array of the form [time,lat,lon], in this order. Mainly for PCA analysis and grid reconstruction.

Usage

mat2Dto3Darray.stations(mat2D, x, y)

Arguments

mat2D
A 2D matrix with time in rows and lonlat in columns, as returned by array3Dto2Dmat.stations

x
unique X coordinates of the points, in ascending order

y
As argument x, for the Y coordinates

Details

The function is the inverse of array3Dto2Dmat.stations

Value

A 3-dimensional array with the dimensions ordered: [time,lat,lon]

Author(s)

M. Iturbide

See Also

array3Dto2Dmat, which performs the inverse operation.

Other internal.helpers: array3Dto2Dmat.stations, array3Dto2Dmat, bindGrid.member, bindGrid.spatial, bindGrid.time, bindGrid, grid3Dto2Dmatrix, redim, sortDim.spatial, sortDim.time

[Package transformeR version 1.4.2 Index]