memberAggregation {transformeR}	R Documentation

Member aggregation

Description

Aggregate a grid along its member dimension

Usage

memberAggregation(grid, aggr.mem, parallel, max.ncores, ncores)

Arguments

grid
A multimember grid to apply the aggregation

aggr.mem
Character string indicatins the aggregation function

parallel.pars
Arguments defining the parallelization options, as passed by parallelCheck

Details

The function preserves the metadadata associated with member information (i.e. initialization dates and member names). In addition, an attribute indicating the member aggregation function is added to the Variable component.

Author(s)

J. Bedia

[Package transformeR version 1.4.2 Index]