overGrid {transformeR}	R Documentation

Spatial overlay for grids and Spatial* objects

Description

Application of function over from sp to grids or station data.

Usage

overGrid(grid, layer, subset = FALSE)

Arguments

grid
Input grid or station data.

layer
SpatialPolygons object from which the geometries or attributes are queried (type help(package = "sp").

subset
Logical (default is FALSE). If TRUE, spatial subsetting is performed over the imput grid using the bounding coordinates of argument 'layer'.

Details

All grid locations outside layer are filled with NAs.

Value

A grid or station data.

A grid

Author(s)

M. Iturbide

See Also

Other subsetting: getTemporalIntersection, intersectGrid.spatial, intersectGrid.time, intersectGrid, subsetDimension, subsetGrid, subsetMembers, subsetRuntime, subsetSeason, subsetSpatial, subsetStation, subsetVar, subsetYears

[Package transformeR version 1.4.2 Index]