prinComp.scale {transformeR}	R Documentation

Local/Global grid scaling

Description

Scale a grid prior to PCA analysis (internal of prinComp)

Usage

prinComp.scale(var.list, scaling)

Arguments

var.list
A named nested list (variables-members) of fields in the form of 2D matrices (after array3Dto2Dmat)

scaling
Character vector of scaling types (same length as var.list)

Details

The "gridbox" scaling performs the scaling independently for each grid/station point (the mean and sd is calculated sepparately for each point). On the contrary, the "field" approach substracts the same global mean and divides by the global sd considering the whole field.

Value

A named list of the same structure as the input var.list with the scaled (and centered) fields (the names of the elements of the list are the variable names).

Author(s)

J Bedia

[Package transformeR version 1.4.2 Index]