sortDim.spatial {transformeR}	R Documentation

Grid sorting by latitude or longitude

Description

Returns a grid with time sorted in ascending order. For internal use mainly, the function is useful for instance when subsetting along time and binding with bindGrid.time, so the user does not need to worry about the ordering of the input.

Usage

sortDim.spatial(grid, dimension = c("y", "x"))

Arguments

grid
Input grid

dimension
longitude ("x") or latitude ("y")

Value

A grid fo similar characteristics, but with time dimension re-arranged in ascending ordered if needed.

Author(s)

M Iturbide

See Also

Other internal.helpers: array3Dto2Dmat.stations, array3Dto2Dmat, bindGrid.member, bindGrid.spatial, bindGrid.time, bindGrid, grid3Dto2Dmatrix, mat2Dto3Darray.stations, redim, sortDim.time

[Package transformeR version 1.4.2 Index]