subsetDecadalForecast {transformeR}	R Documentation

Subset decadal forecast by initialization years

Description

Subset decadal forecast by initialization years

Usage

subsetDecadalForecast(data, ly = c(1, 2))

Arguments

data
grid obtained from the output of function loadDecadalForecast.

ly
Vector of integers indicating the lead years of initialization. Default is ly = c(1,2) (the two years before).

Value

A grid object with an additional dimension corresponding to the selected lead years.

Author(s)

M. Iturbide

[Package transformeR version 1.4.2 Index]