subsetRuntime {transformeR}	R Documentation

Runtime subsets from a grid

Description

Retrieves a grid that is a logical subset of a multiruntime grid along its 'runtime' dimension. Multiruntime multigrids are supported. Subroutine of subsetGrid.

Usage

subsetRuntime(grid, runtime)

Arguments

grid
Input multiruntime grid to be subset.

runtime
An integer vector indicating the position of the runtimes to be subset.

Details

An attribute 'subset' with value 'subsetRuntime' is added to the Runtime slot of the output subset.

Value

A grid (or multigrid) that is a logical subset of the input grid along its 'runtime' dimension.

Author(s)

M. Iturbide

See Also

Other subsetting: getTemporalIntersection, intersectGrid.spatial, intersectGrid.time, intersectGrid, overGrid, subsetDimension, subsetGrid, subsetMembers, subsetSeason, subsetSpatial, subsetStation, subsetVar, subsetYears

[Package transformeR version 1.4.2 Index]