subsetSeason {transformeR}	R Documentation

Monthly subset of a grid

Description

Retrieves a grid that is a logical subset of the input grid along its 'time' dimension, on a monthly basis. Multimember multigrids are supported. Subroutine of subsetGrid.

Usage

subsetSeason(grid, season = NULL)

Arguments

grid
Input grid to be subset (possibly a multimember/multigrid).

season
An integer vector indicating the months to be subset.

Details

An attribute 'subset' with value 'time' is added to the Variable slot of the output grid.

Value

A grid (or multigrid) that is a logical subset of the input grid along its 'time' dimension.

Author(s)

J. Bedia

See Also

Other subsetting: getTemporalIntersection, intersectGrid.spatial, intersectGrid.time, intersectGrid, overGrid, subsetDimension, subsetGrid, subsetMembers, subsetRuntime, subsetSpatial, subsetStation, subsetVar, subsetYears

[Package transformeR version 1.4.2 Index]