subsetSpatial {transformeR}	R Documentation

Spatial subset from a grid

Description

Retrieves a grid that is a logical subset of the input grid along its 'lat' and 'lon' dimensions. Multimember multigrids are supported. Subroutine of subsetGrid.

Usage

subsetSpatial(grid, lonLim, latLim, outside)

Arguments

grid
Input grid to be subset (possibly a multimember multigrid).

lonLim
Vector of length = 2, with minimum and maximum longitude coordinates, in decimal degrees, of the bounding box defining the subset. For single-point subsets, a numeric value with the longitude coordinate. If NULL (default), no subsetting is performed on the longitude dimension

latLim
Same as lonLim argument, but for latitude.

outside
Logical. Default to FALSE. If TRUE, subset coordinates outside the grid extent are allowed.

Details

An attribute subset with value subsetSpatial is added to the xyCoords component of the output grid.

Value

A grid (or multigrid) that is a logical spatial subset of the input grid.

Author(s)

J. Bedia

See Also

Other subsetting: getTemporalIntersection, intersectGrid.spatial, intersectGrid.time, intersectGrid, overGrid, subsetDimension, subsetGrid, subsetMembers, subsetRuntime, subsetSeason, subsetStation, subsetVar, subsetYears

[Package transformeR version 1.4.2 Index]