typeofGrid {transformeR}	R Documentation

Type of grid

Description

Describe the type of grid according to the spatial representation of the data

Usage

typeofGrid(grid)

Arguments

grid
An input grid

Value

A character vector decribing the type of spatial reference. Three disjoint values possible: "station", "rotated_grid" and "regular_grid"

Author(s)

J Bedia

See Also

Other get.helpers: get2DmatCoordinates, getCoordinates, getDim, getGridVerticalLevels, getGrid, getRefDates, getSeason, getShape, getTimeResolution, getVarNames, getYearsAsINDEX, which.leap

[Package transformeR version 1.4.2 Index]