which.leap {transformeR}	R Documentation

Leap years

Description

Return the indices of leap years in a vector of years

Usage

which.leap(years)

Arguments

years
a integer vector of (gregorian) years

Value

a vector of indices of the position of leap years

Author(s)

J. Bedia

References

https://en.wikipedia.org/wiki/Leap_year#Algorithm

See Also

Other get.helpers: get2DmatCoordinates, getCoordinates, getDim, getGridVerticalLevels, getGrid, getRefDates, getSeason, getShape, getTimeResolution, getVarNames, getYearsAsINDEX, typeofGrid

Examples

leap.years <- which.leap(1885:1937)
(1885:1937)[leap.years]

[Package transformeR version 1.4.2 Index]